server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
alembic
python-multipart
email-validator
